fix:ゲーム画面の設定部分をgameCanvasConfig.jsに移動,gameidをformに
