+++ frontend/utils/locales/ja/gameplay.json
@@ -9,8 +9,7 @@
   },
   "Remote_OFF": "リモートオフ",
   "Remote_ON": "リモートオン",
-  "enter_id": "ゲームIDを入力",
-  "set_id": "IDを設定",
+  "enter_gameid": "ゲームIDを入力",
   "set": "設定",
   "error": {
     "start": "ゲームの開始に失敗しました",
