+++ frontend/utils/locales/zh/gameplay.json
@@ -9,8 +9,7 @@
   },
   "Remote_OFF": "遥控关闭",
   "Remote_ON": "遥控开启",
-  "enter_id": "输入您的游戏 ID",
-  "set_id": "设置 ID",
+  "enter_gameid": "输入您的游戏 ID",
   "set": "设置",
   "error": {
     "start": "游戏启动失败",
