+++ frontend/views/templates/Gameplay.html
@@ -21,10 +21,12 @@
   <button type="button" id = "game-start" class="btn btn-success mt-3" data-i18n="gameplay:game_start">Game Start</button>
   <button type="button" id = "remote-mode" class="btn btn-warning mt-3" data-i18n="gameplay:Remote_OFF">Remote OFF</button>
   <div id="remote-options" class="display-none">
-    <div class="d-flex align-items-center">
-      <input type="text" id="game-id-text" class="mt-3 game-id-text-size" data-i18n="gameplay:enter_id" placeholder="Game ID">
-      <button type="button" id="game-id-set-button" class="btn btn-danger ms-2 mt-auto game-id-set-button-size" data-i18n="gameplay:set_id">Set ID</button>
-    </div>
+    <form id="game-id-form">
+      <input id="game-id" type="text" class="form-control mt-3" data-i18n="[placeholder]gameplay:enter_gameid">
+      <div class="submit-button text-center">
+        <button type="submit" class="btn btn-danger mt-3" data-i18n="gameplay:set">Set</button>
+      </div>
+    </form>
     <button type="button" id = "remote-left" class="btn btn-warning mt-3" data-i18n="gameplay:popup.left">Left</button>
     <button type="button" id = "remote-right" class="btn btn-warning mt-3" data-i18n="gameplay:popup.right">Right</button>
     <button type="button" id = "remote-set-button" class="btn btn-danger mt-3" data-i18n="gameplay:set">Set</button>
